@@ -1,8 +1,8 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '5.1.4'
-ruby '2.4.2'
+gem 'rails', '5.2.1'
+ruby '2.5.1'
 
 # Use sqlite3 as the database for Active Record
 
