@@ -1,4 +1,4 @@
-class CreateRelationships < ActiveRecord::Migration
+class CreateRelationships < ActiveRecord::Migration[5.2]
   def change
     create_table :relationships do |t|
       t.integer :follower_id
