@@ -1,4 +1,4 @@
-class CreateContacts < ActiveRecord::Migration
+class CreateContacts < ActiveRecord::Migration[5.2]
   def change
     create_table :contacts do |t|
       t.string :name
