@@ -1,4 +1,4 @@
-class AddConfirmableToDevise < ActiveRecord::Migration
+class AddConfirmableToDevise < ActiveRecord::Migration[5.2]
   # Note: You can't use change, as User.update_all will fail in the down migration
   def up
     add_column :users, :confirmation_token, :string
