@@ -1,4 +1,4 @@
-class AddUnconfirmedemailToDevise < ActiveRecord::Migration
+class AddUnconfirmedemailToDevise < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :unconfirmed_email, :string
   end
