@@ -1,4 +1,4 @@
-class AddSlugToUsers < ActiveRecord::Migration
+class AddSlugToUsers < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :slug, :string
     add_index :users, :slug, unique: true
