@@ -1,4 +1,4 @@
-class AddAvatarToUsers < ActiveRecord::Migration
+class AddAvatarToUsers < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :avatar, :string
   end
