@@ -1,4 +1,4 @@
-class AddColumsToUsers < ActiveRecord::Migration
+class AddColumsToUsers < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :provider, :string
     add_column :users, :uid, :string
