@@ -1,4 +1,4 @@
-class CreateAnswers < ActiveRecord::Migration
+class CreateAnswers < ActiveRecord::Migration[5.2]
   def change
     create_table :answers do |t|
       t.integer :answering_id
