@@ -1,4 +1,4 @@
-class AddDescriptionToGroups < ActiveRecord::Migration
+class AddDescriptionToGroups < ActiveRecord::Migration[5.2]
   def change
     add_column :groups, :description, :text
   end
