@@ -1,4 +1,4 @@
-class CreateInvites < ActiveRecord::Migration
+class CreateInvites < ActiveRecord::Migration[5.2]
   def change
     create_table :invites do |t|
       t.string :email
