@@ -1,4 +1,4 @@
-class CreatePhotoGalleries < ActiveRecord::Migration
+class CreatePhotoGalleries < ActiveRecord::Migration[5.2]
   def change
     create_table :photo_galleries do |t|
       t.string :title
