@@ -1,4 +1,4 @@
-class AddPhotoGalleryIdToPhotos < ActiveRecord::Migration
+class AddPhotoGalleryIdToPhotos < ActiveRecord::Migration[5.2]
   def change
     add_column :photos, :photo_gallery_id, :integer
   end
