@@ -1,4 +1,4 @@
-class CreateNotices < ActiveRecord::Migration
+class CreateNotices < ActiveRecord::Migration[5.2]
   def change
     create_table :notices do |t|
       t.text :description
