@@ -1,4 +1,4 @@
-class CreateJoinTableNoticesPhotos < ActiveRecord::Migration
+class CreateJoinTableNoticesPhotos < ActiveRecord::Migration[5.2]
   def change
     create_join_table :notices, :photos do |t|
       t.index [:notice_id, :photo_id]
