@@ -1,4 +1,4 @@
-class CreatePhotoComments < ActiveRecord::Migration
+class CreatePhotoComments < ActiveRecord::Migration[5.2]
   def change
     create_table :photo_comments do |t|
       t.integer :photo_id
