@@ -1,4 +1,4 @@
-class AddUserToPhoto < ActiveRecord::Migration
+class AddUserToPhoto < ActiveRecord::Migration[5.2]
   def change
     add_column :photos, :user_id, :integer
   end
