@@ -1,4 +1,4 @@
-class AddColumnToEmailNotifications < ActiveRecord::Migration
+class AddColumnToEmailNotifications < ActiveRecord::Migration[5.2]
   def change
     add_column :email_notifications, :no_emails, :boolean
   end
