@@ -1,4 +1,4 @@
-class RemoveGetItFromEmailNotification < ActiveRecord::Migration
+class RemoveGetItFromEmailNotification < ActiveRecord::Migration[5.2]
   def change
     remove_column :email_notifications, :get_it
   end
