@@ -1,4 +1,4 @@
-class CreateCustomAppearances < ActiveRecord::Migration[5.1]
+class CreateCustomAppearances < ActiveRecord::Migration[5.2]
   def change
     create_table :custom_appearances do |t|
       t.string :navbar
