@@ -1,4 +1,4 @@
-class AddLinkcolorToCustomAppearance < ActiveRecord::Migration[5.1]
+class AddLinkcolorToCustomAppearance < ActiveRecord::Migration[5.2]
   def change
     add_column :custom_appearances, :linkcolor, :text
   end
