@@ -1,4 +1,4 @@
-class AddGroupIdToReadPosts < ActiveRecord::Migration[5.1]
+class AddGroupIdToReadPosts < ActiveRecord::Migration[5.2]
   def change
     add_column :read_posts, :group_id, :integer
   end
