@@ -1,4 +1,4 @@
-class AddPrivateToGroups < ActiveRecord::Migration[5.1]
+class AddPrivateToGroups < ActiveRecord::Migration[5.2]
   def change
     add_column :groups, :private, :boolean
   end
