@@ -10,32 +10,35 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180410055630) do
+ActiveRecord::Schema.define(version: 2018_04_10_055630) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "activities", id: :serial, force: :cascade do |t|
-    t.integer "trackable_id"
-    t.string "trackable_type", limit: 255
-    t.integer "owner_id"
-    t.string "owner_type", limit: 255
-    t.string "key", limit: 255
+  create_table "activities", force: :cascade do |t|
+    t.string "trackable_type"
+    t.bigint "trackable_id"
+    t.string "owner_type"
+    t.bigint "owner_id"
+    t.string "key"
     t.text "parameters"
-    t.integer "recipient_id"
-    t.string "recipient_type", limit: 255
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.string "recipient_type"
+    t.bigint "recipient_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["owner_id", "owner_type"], name: "index_activities_on_owner_id_and_owner_type"
+    t.index ["owner_type", "owner_id"], name: "index_activities_on_owner_type_and_owner_id"
     t.index ["recipient_id", "recipient_type"], name: "index_activities_on_recipient_id_and_recipient_type"
+    t.index ["recipient_type", "recipient_id"], name: "index_activities_on_recipient_type_and_recipient_id"
     t.index ["trackable_id", "trackable_type"], name: "index_activities_on_trackable_id_and_trackable_type"
+    t.index ["trackable_type", "trackable_id"], name: "index_activities_on_trackable_type_and_trackable_id"
   end
 
-  create_table "answers", id: :serial, force: :cascade do |t|
+  create_table "answers", force: :cascade do |t|
     t.integer "comment_id"
     t.integer "micropost_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "blogrolls", force: :cascade do |t|
@@ -46,28 +49,28 @@ ActiveRecord::Schema.define(version: 20180410055630) do
     t.integer "custom_appearance_id"
   end
 
-  create_table "comments", id: :serial, force: :cascade do |t|
+  create_table "comments", force: :cascade do |t|
     t.text "body"
-    t.integer "user_id"
+    t.bigint "user_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["user_id", "created_at"], name: "index_comments_on_user_id_and_created_at"
     t.index ["user_id"], name: "index_comments_on_user_id"
   end
 
-  create_table "contacts", id: :serial, force: :cascade do |t|
-    t.string "name", limit: 255
-    t.string "email", limit: 255
+  create_table "contacts", force: :cascade do |t|
+    t.string "name"
+    t.string "email"
     t.text "message"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "conversations", id: :serial, force: :cascade do |t|
+  create_table "conversations", force: :cascade do |t|
     t.integer "sender_id"
     t.integer "recipient_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["recipient_id"], name: "index_conversations_on_recipient_id"
     t.index ["sender_id"], name: "index_conversations_on_sender_id"
   end
@@ -82,17 +85,17 @@ ActiveRecord::Schema.define(version: 20180410055630) do
     t.index ["user_id"], name: "index_custom_appearances_on_user_id", unique: true
   end
 
-  create_table "email_notifications", id: :serial, force: :cascade do |t|
+  create_table "email_notifications", force: :cascade do |t|
     t.integer "user_id"
     t.integer "micropost_id"
     t.boolean "emails"
   end
 
-  create_table "friendly_id_slugs", id: :serial, force: :cascade do |t|
-    t.string "slug", limit: 255, null: false
+  create_table "friendly_id_slugs", force: :cascade do |t|
+    t.string "slug", null: false
     t.integer "sluggable_id", null: false
     t.string "sluggable_type", limit: 50
-    t.string "scope", limit: 255
+    t.string "scope"
     t.datetime "created_at"
     t.index ["slug", "sluggable_type", "scope"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type_and_scope", unique: true
     t.index ["slug", "sluggable_type"], name: "index_friendly_id_slugs_on_slug_and_sluggable_type"
@@ -100,94 +103,96 @@ ActiveRecord::Schema.define(version: 20180410055630) do
     t.index ["sluggable_type"], name: "index_friendly_id_slugs_on_sluggable_type"
   end
 
-  create_table "group_maintainers", id: :serial, force: :cascade do |t|
+  create_table "group_maintainers", force: :cascade do |t|
     t.integer "group_id"
     t.integer "user_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "group_memberships", id: :serial, force: :cascade do |t|
+  create_table "group_memberships", force: :cascade do |t|
     t.integer "user_id"
     t.integer "group_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["user_id", "group_id"], name: "index_group_memberships_on_user_id_and_group_id", unique: true
   end
 
-  create_table "groups", id: :serial, force: :cascade do |t|
-    t.string "name", limit: 255
-    t.datetime "created_at"
-    t.datetime "updated_at"
+  create_table "groups", force: :cascade do |t|
+    t.string "name"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.text "description"
     t.boolean "private"
   end
 
-  create_table "invite_counts", id: :serial, force: :cascade do |t|
+  create_table "invite_counts", force: :cascade do |t|
     t.integer "invite_id"
     t.integer "user_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.boolean "deleted", default: false
   end
 
-  create_table "invites", id: :serial, force: :cascade do |t|
-    t.string "email", limit: 255
-    t.datetime "created_at"
-    t.datetime "updated_at"
+  create_table "invites", force: :cascade do |t|
+    t.string "email"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.text "letter"
     t.index ["email"], name: "index_invites_on_email", unique: true
   end
 
-  create_table "likes", id: :serial, force: :cascade do |t|
-    t.string "liker_type", limit: 255
+  create_table "likes", force: :cascade do |t|
+    t.string "liker_type"
     t.integer "liker_id"
-    t.string "likeable_type", limit: 255
+    t.string "likeable_type"
     t.integer "likeable_id"
     t.datetime "created_at"
     t.index ["likeable_id", "likeable_type"], name: "fk_likeables"
     t.index ["liker_id", "liker_type"], name: "fk_likes"
   end
 
-  create_table "mailboxer_conversation_opt_outs", id: :serial, force: :cascade do |t|
-    t.integer "unsubscriber_id"
-    t.string "unsubscriber_type", limit: 255
-    t.integer "conversation_id"
+  create_table "mailboxer_conversation_opt_outs", force: :cascade do |t|
+    t.string "unsubscriber_type"
+    t.bigint "unsubscriber_id"
+    t.bigint "conversation_id"
     t.index ["conversation_id"], name: "index_mailboxer_conversation_opt_outs_on_conversation_id"
-    t.index ["unsubscriber_id", "unsubscriber_type"], name: "index_mailboxer_conversation_opt_outs_on_unsubscriber_id_type"
+    t.index ["unsubscriber_type", "unsubscriber_id"], name: "conversation_unsubscr_index"
   end
 
-  create_table "mailboxer_conversations", id: :serial, force: :cascade do |t|
-    t.string "subject", limit: 255, default: ""
+  create_table "mailboxer_conversations", force: :cascade do |t|
+    t.string "subject", default: ""
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "mailboxer_notifications", id: :serial, force: :cascade do |t|
-    t.string "type", limit: 255
+  create_table "mailboxer_notifications", force: :cascade do |t|
+    t.string "type"
     t.text "body"
-    t.string "subject", limit: 255, default: ""
-    t.integer "sender_id"
-    t.string "sender_type", limit: 255
+    t.string "subject", default: ""
+    t.string "sender_type"
+    t.bigint "sender_id"
     t.integer "conversation_id"
     t.boolean "draft", default: false
-    t.string "notification_code", limit: 255
-    t.integer "notified_object_id"
-    t.string "notified_object_type", limit: 255
-    t.string "attachment", limit: 255
+    t.text "notification_code"
+    t.string "notified_object_type"
+    t.bigint "notified_object_id"
+    t.string "attachment"
     t.datetime "updated_at", null: false
     t.datetime "created_at", null: false
     t.boolean "global", default: false
     t.datetime "expires"
     t.index ["conversation_id"], name: "index_mailboxer_notifications_on_conversation_id"
     t.index ["notified_object_id", "notified_object_type"], name: "index_mailboxer_notifications_on_notified_object_id_and_type"
+    t.index ["notified_object_type", "notified_object_id"], name: "mailboxer_notifications_notified_object"
     t.index ["sender_id", "sender_type"], name: "index_mailboxer_notifications_on_sender_id_and_sender_type"
+    t.index ["sender_type", "sender_id"], name: "index_mailboxer_notifications_on_sender_type_and_sender_id"
     t.index ["type"], name: "index_mailboxer_notifications_on_type"
   end
 
-  create_table "mailboxer_receipts", id: :serial, force: :cascade do |t|
-    t.integer "receiver_id"
-    t.string "receiver_type", limit: 255
+  create_table "mailboxer_receipts", force: :cascade do |t|
+    t.string "receiver_type"
+    t.bigint "receiver_id"
     t.integer "notification_id", null: false
     t.boolean "is_read", default: false
     t.boolean "trashed", default: false
@@ -200,92 +205,93 @@ ActiveRecord::Schema.define(version: 20180410055630) do
     t.string "message_id"
     t.index ["notification_id"], name: "index_mailboxer_receipts_on_notification_id"
     t.index ["receiver_id", "receiver_type"], name: "index_mailboxer_receipts_on_receiver_id_and_receiver_type"
+    t.index ["receiver_type", "receiver_id"], name: "index_mailboxer_receipts_on_receiver_type_and_receiver_id"
   end
 
-  create_table "mentions", id: :serial, force: :cascade do |t|
-    t.string "mentioner_type", limit: 255
+  create_table "mentions", force: :cascade do |t|
+    t.string "mentioner_type"
     t.integer "mentioner_id"
-    t.string "mentionable_type", limit: 255
+    t.string "mentionable_type"
     t.integer "mentionable_id"
     t.datetime "created_at"
     t.index ["mentionable_id", "mentionable_type"], name: "fk_mentionables"
     t.index ["mentioner_id", "mentioner_type"], name: "fk_mentions"
   end
 
-  create_table "messages", id: :serial, force: :cascade do |t|
+  create_table "messages", force: :cascade do |t|
     t.text "body"
-    t.integer "conversation_id"
-    t.integer "user_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.bigint "conversation_id"
+    t.bigint "user_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["conversation_id"], name: "index_messages_on_conversation_id"
     t.index ["user_id"], name: "index_messages_on_user_id"
   end
 
-  create_table "microposts", id: :serial, force: :cascade do |t|
+  create_table "microposts", force: :cascade do |t|
     t.text "content"
-    t.integer "user_id"
+    t.bigint "user_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.string "picture", limit: 255
+    t.string "picture"
     t.integer "group_id"
     t.index ["user_id", "created_at"], name: "index_microposts_on_user_id_and_created_at"
     t.index ["user_id"], name: "index_microposts_on_user_id"
   end
 
-  create_table "notices", id: :serial, force: :cascade do |t|
+  create_table "notices", force: :cascade do |t|
     t.text "description"
     t.integer "user_id"
-    t.string "picture", limit: 255
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.string "picture"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer "photo_gallery_id"
     t.integer "count", default: 0
   end
 
   create_table "notices_photos", id: false, force: :cascade do |t|
-    t.integer "notice_id", null: false
-    t.integer "photo_id", null: false
+    t.bigint "notice_id", null: false
+    t.bigint "photo_id", null: false
     t.index ["notice_id", "photo_id"], name: "index_notices_photos_on_notice_id_and_photo_id"
     t.index ["photo_id", "notice_id"], name: "index_notices_photos_on_photo_id_and_notice_id"
   end
 
-  create_table "pg_search_documents", id: :serial, force: :cascade do |t|
+  create_table "pg_search_documents", force: :cascade do |t|
     t.text "content"
-    t.integer "searchable_id"
-    t.string "searchable_type", limit: 255
+    t.string "searchable_type"
+    t.bigint "searchable_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.index ["searchable_id", "searchable_type"], name: "index_pg_search_documents_on_searchable_id_and_searchable_type"
+    t.index ["searchable_type", "searchable_id"], name: "index_pg_search_documents_on_searchable_type_and_searchable_id"
   end
 
-  create_table "photo_comments", id: :serial, force: :cascade do |t|
+  create_table "photo_comments", force: :cascade do |t|
     t.integer "photo_id"
     t.integer "comment_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
-  create_table "photo_galleries", id: :serial, force: :cascade do |t|
-    t.string "title", limit: 255
+  create_table "photo_galleries", force: :cascade do |t|
+    t.string "title"
     t.text "description"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer "last_updated_by_id"
   end
 
   create_table "photo_galleries_users", id: false, force: :cascade do |t|
-    t.integer "photo_gallery_id", null: false
-    t.integer "user_id", null: false
+    t.bigint "photo_gallery_id", null: false
+    t.bigint "user_id", null: false
     t.index ["photo_gallery_id", "user_id"], name: "index_photo_galleries_users_on_photo_gallery_id_and_user_id"
     t.index ["user_id", "photo_gallery_id"], name: "index_photo_galleries_users_on_user_id_and_photo_gallery_id"
   end
 
-  create_table "photos", id: :serial, force: :cascade do |t|
+  create_table "photos", force: :cascade do |t|
     t.integer "micropost_id"
-    t.string "picture", limit: 255
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.string "picture"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.integer "photo_gallery_id"
     t.integer "user_id"
   end
@@ -299,52 +305,53 @@ ActiveRecord::Schema.define(version: 20180410055630) do
     t.integer "group_id"
   end
 
-  create_table "relationships", id: :serial, force: :cascade do |t|
+  create_table "relationships", force: :cascade do |t|
     t.integer "follower_id"
     t.integer "followed_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
     t.index ["followed_id"], name: "index_relationships_on_followed_id"
     t.index ["follower_id", "followed_id"], name: "index_relationships_on_follower_id_and_followed_id", unique: true
     t.index ["follower_id"], name: "index_relationships_on_follower_id"
   end
 
-  create_table "users", id: :serial, force: :cascade do |t|
-    t.string "email", limit: 255, default: "", null: false
-    t.string "encrypted_password", limit: 255, default: "", null: false
-    t.string "reset_password_token", limit: 255
+  create_table "users", force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
     t.integer "sign_in_count", default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string "current_sign_in_ip", limit: 255
-    t.string "last_sign_in_ip", limit: 255
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.string "name", limit: 255
-    t.string "confirmation_token", limit: 255
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "name"
+    t.string "confirmation_token"
     t.datetime "confirmed_at"
     t.datetime "confirmation_sent_at"
-    t.string "unconfirmed_email", limit: 255
-    t.string "slug", limit: 255
-    t.string "avatar", limit: 255
-    t.string "provider", limit: 255
-    t.string "uid", limit: 255
+    t.string "unconfirmed_email"
+    t.string "slug"
+    t.string "avatar"
+    t.string "provider"
+    t.string "uid"
     t.boolean "admin", default: false
-    t.string "invitation_token", limit: 255
+    t.string "invitation_token"
     t.datetime "invitation_created_at"
     t.datetime "invitation_sent_at"
     t.datetime "invitation_accepted_at"
     t.integer "invitation_limit"
-    t.integer "invited_by_id"
-    t.string "invited_by_type", limit: 255
+    t.string "invited_by_type"
+    t.bigint "invited_by_id"
     t.integer "invitations_count", default: 0
     t.index ["confirmation_token"], name: "index_users_on_confirmation_token", unique: true
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["invitation_token"], name: "index_users_on_invitation_token", unique: true
     t.index ["invitations_count"], name: "index_users_on_invitations_count"
     t.index ["invited_by_id"], name: "index_users_on_invited_by_id"
+    t.index ["invited_by_type", "invited_by_id"], name: "index_users_on_invited_by_type_and_invited_by_id"
     t.index ["name"], name: "index_users_on_name", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
     t.index ["slug"], name: "index_users_on_slug", unique: true
